Change names of protocol a bit
